do some rework
